Build: Fix compilation on VS2017

+++ Library/OcAfterBootCompatLib/ServiceOverrides.c
@@ -351,7 +351,7 @@ ApplyBooterPatch (
     Patch->Replace,
     Patch->ReplaceMask,
     ImageBase,
-    ImageSize,
+    (UINT32) ImageSize,
     Patch->Count,
     Patch->Skip
     );
